requirements.txt
Django==5.1.3
mysqlclient==2.2.1
python-dotenv==1.0.0
requests==2.31.0
paypalcheckoutsdk==1.0.1
Pillow==10.1.0
django-paypal==2.0
python-dateutil==2.8.2
pytz==2023.3.post1
whitenoise==6.6.0
gunicorn==21.2.0
geoip2==4.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
